feat: moved cyberconnect types
